Bad Performace
The bad performance is due to the bad labelling process and training process. This image was not well processed to extract all the samples and their labels from it. Fix it later. Or choose some new training sets (Remember to change the rectangle sizes due to the different size of pictures)